QMessageLogContext: rename and improve copy()
The "copy" name led me to mis-believe that it was a workaround
to the class being non-copiable, and was missing copying a
data member.

This is instead deliberate; so rename the function to
"copyContextFrom", documenting that version won't be copied.
The function is still private API, so actually make it private;
QMessageLogContext already befriended the only user.

Finally, make it return *this (so that it can be chained,
if necessary) and noexcept.

Change-Id: I3deb3c8edc863a88ac0c37467b144ec2e20db5ca
Reviewed-by: Marc Mutz <marc.mutz@kdab.com>
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>